Reorganize common/, refs #253
