+++ src/pretalx/agenda/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.utils.translation import ugettext as _
 
-from pretalx.common.forms import ReadOnlyFlag
+from pretalx.common.mixins.forms import ReadOnlyFlag
 from pretalx.submission.models import Feedback
