+++ src/pretalx/common/forms/__init__.py
@@ -1,0 +1,6 @@
+from .forms import I18nFormSet, SearchForm
+
+__all__ = [
+    'I18nFormSet',
+    'SearchForm',
+]
