+++ src/pretalx/common/mixins/models.py
