+++ src/pretalx/orga/forms/cfp.py
@@ -3,7 +3,7 @@ from django.utils.translation import ugettext_lazy as _
 from hierarkey.forms import HierarkeyForm
 from i18nfield.forms import I18nFormMixin, I18nModelForm
 
-from pretalx.common.forms import ReadOnlyFlag
+from pretalx.common.mixins.forms import ReadOnlyFlag
 from pretalx.common.phrases import phrases
 from pretalx.submission.models import (
     AnswerOption, CfP, Question, SubmissionType,
